Fix 'chorus' bug.
When a playing piano was paused, and then played again, it would cause the notes to be 'layered', because a new synth would be created and played along with the original. Fixed by separating synth creation from synth control.

Also tried to move some functionality to the 'init' button, like synth creation and calling Tone.start() to try to fix the latency bug when pressing play, but it didn't work.